requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioop-lts==0.2.1
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
edge-tts==7.0.1
espeakng==1.0.3
frozenlist==1.6.0
greenlet==3.2.1
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.24
langchain-core==0.3.56
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langchain-together==0.3.0
langsmith==0.3.37
multidict==6.4.3
openai==1.76.0
orjson==3.10.16
packaging==24.2
propcache==0.3.1
PyAudio==0.2.14
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
pyttsx3==2.98
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SpeechRecognition==3.14.2
SQLAlchemy==2.0.40
srt==3.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
zstandard==0.23.0
